+++ packages/cli/src/zosfiles/-strings-/en.ts
@@ -830,7 +830,10 @@ export default {
                 "The maximum number of TSO address spaces have been created. When you specify 0, " +
                 Constants.DISPLAY_NAME + " attempts to upload all members at once" +
                 " without a maximum number of concurrent requests. ",
-            INCLUDE_HIDDEN: "Include hidden files and folders that have names beginning with a dot."
+            INCLUDE_HIDDEN: "Include hidden files and folders that have names beginning with a dot.",
+            MAKE_DIR: "Create the target directory if it does not already exist.",
+            SKIP_DIRECTORY_CHECK: "Specifies whether to skip checking if the target path is a directory before uploading. " +
+             "When set to true, the upload will proceed assuming the target is a file path."
         }
     },
     VIEW: {